LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/declarative-shadow-dom-slot-assignment-serialization.tentative-expected.txt


PASS shadowrootslotassignment=manual is serialized and appears before shadowrootclonable and shadowrootserializable
PASS shadowrootslotassignment=named is not serialized as it's the default
PASS shadowrootslotassignment=manual serializes between shadowrootdelegatesfocus and shadowrootclonable
